Move setup_amqp and setup_database to reactive handler file
They don't have any interaction with barbican charm class code so they
make better sense in the handlers file.
